Add files via upload
Array Class
Using classes and arrays, the team will develop a class MyArray that holds an integer 
array. The class will have:
1- A dynamic integer array
2- The size of the array
3- A constructor that takes the arrays size as a parameter and initializes array 
elements to 0
4- A copy constructor
5- A getter to the array size; getSize()
6- Overload the subscript operator [ ] to get and set array elements
7- Overload the assignment operator=.
8- Overload the extraction operator << to display the array elements.
9- Overload the operator + to concatenate 2 arrays and return the result in a third 
array object. Example : MyArray L3 = L1 + L2.
Note: the arrays L1 and L2 will not be changed, only L3 has the result of both
of their elements concatenated one after the other.
10- Overload the operator - to return an object that has the array elements of 
the first object excluding the array elements of the second object. 
Example : MyArray L3 = L1 - L2. //if array of L1 is [3,6,9,1,8] and array of L2 
is [1,3], then L3 will have the array [6,9,8]
Note: the arrays L1 and L2 will not be changed, only L3 is returned.
11- Overload the operator * that multiplies the array elements by an integer 
number and return the result in another MyArray object. Ex. L2 = L1 * 10;
12- Add a member function resize that expands the array size. The function 
has an integer parameter with the new size. If the parameter value is less than 
the current size return false, otherwise reallocate the array (preserving already 
existing elements) to the new given size and return true.
Note that the extra elements are initialized to zero in this function
13- Add a destructor to your class.
####################################################
1- Create an array A
2- Enter array elements of A
3- Copy array A into a new array and display the new array
4- Display an element in array A by index
5- Set an element in array A by index
6- Display array A
7- Concatenate 2 arrays into array A
8- Subtract 2 arrays into array A
9- Multiply array elements of A by an integer
10- Expand array A
11- Get size of array A
12- Exit