Add files via upload
A class Player is a class that holds the info of a football player including the following 
private member variables:
1- Name (string)
2- Year of Birth (int)
3- Height (int)
4- International Goals (int)
5- Nationality
The class NationalTeam is a class holding the players in a national football team of a 
certain country. The class has an array of Player objects [players] of size 18, holding 
all the players including substitutes (fixed size). 
The class has another array of Booleans [active] of size 18, where an element at index i 
in the active array is true if there’s a player in the players array at index i, and false 
otherwise. Both arrays are private members of the class.
Other members of the class include country.
The array of players is initially empty, and the active array’s elements are set to false. 
We can add to the array using the addPlayer(const Player&) member function, which 
will add the parameter player object in the first empty slot in the players array. The 
function will also update the active array to indicate that a player is added at that 
index. The function returns true if the player was added, false otherwise.
We can also remove players using removePlayer(string) from the players array by 
giving the player’s name and the player will be removed (the player object will not 
actually be removed but, the active array will be updated to indicate this). Make sure 
that the name entered is not case sensitive. The function will return true if the player 
was found and removed, false otherwise.
There other public member functions in the class including: 
- getNumOfPlayers // to return total number of players in the team
- AvgHeight() // to return the average height of the team players
- OldestPlayer() // to return the oldest player object 
- YoungestPlayer() // to return the youngest player object
- AvgIntGoals() // to return the average number of international goals per 
player in the team
- highestIntGoals() // to return the player with the highest number of 
international goals 
- operator<< // to output a team’s country and its players info.
- writeStatistics(oftream&)
// writes team statistics to a file including:
o Team Country
o Number of players
o Average height of players
o Name and age of oldest player
o Name and age of youngest player
o Average number of international goals of players
Name of the player with the highest international goals and the number 
of goals scored by this player.
Ex.
Create a global function readPlayersFromFile(ifstream&) which reads all players 
info from a file named “players.txt” into an array of player objects. Note that this file 
will have all of the players of all the teams that will be added later. The file will have 
the number of players in the file at the first line of the file.
An example file is shown:
Finally, create a main that will:
1- read players from players.txt into an array of player objects.
2- Ask the user to enter a number of national teams where the user will just enter their 
country names.
3- Next Display the players read from the file and the national teams entered by the 
user.
4- Now in a loop:
Ask the user to add/remove players to teams as many times as they wish
You can display players and teams numbered in step 3 such that the user can choose 
to add a player to a team by their displayed numbers.
To remove a player the user has to enter their name to be removed. 
5- After finishing step 4, show all of the national teams added by the user including 
the country and the players in this team using the operator<<.
6- Finally write the statistics of each of the teams to the file statistics.txt. All teams 
statistics should be in this single file.